requirements.txt
asgiref==3.8.1
Django==5.1.4
django-extensions==3.2.3
PyYAML==6.0.2
sqlparse==0.5.3
pre-commit==4.0.1
django-colorfield==0.11.0
djangorestframework==3.15.2
pytest==8.3.4
pytest-django==4.9.0
factory_boy==3.3.0
django-filter==24.3
